url_shortener/requirements.txt
Flask-SQLAlchemy==2.5.0
